pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pyrogram>=2.0.106",
    "python-telegram-bot==21.5",
    "telethon>=1.41.2",
]
